pyproject.toml
[project]
name = "custom-scripts"
version = "0.1.0"
description = "Custom scripts and utilities"
requires-python = ">=3.11"
dependencies = [
    "pyxdg>=0.28",
]

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["helpers"]
